LayoutTests/imported/w3c/web-platform-tests/css/css-view-transitions/navigation/pageswap-push-with-redirect-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_array_equals: incorrect event order lengths differ, expected array ["pageswap", "transition", "http://localhost:8800/css/css-view-transitions/navigation/pageswap-push-with-redirect.html?...

Harness Error (FAIL), message = Unhandled rejection: assert_array_equals: incorrect event order lengths differ, expected array ["pageswap", "transition", "http://localhost:8800/css/css-view-transitions/navigation/pageswap-push-with-redirect.html?new", "push", "from", "pagehide"] length 6, got ["pagehide"] length 1

TIMEOUT pageswap on navigation with same-origin redirect Test timed out

Harness Error (FAIL), message = Unhandled rejection: assert_array_equals: incorrect event order lengths differ, expected array ["pageswap", "transition", "http://localhost:8800/css/css-view-transitions/navigation/pageswap-push-with-redirect.html?new", "push", "from", "pagehide"] length 6, got ["pagehide"] length 1

TIMEOUT pageswap on navigation with same-origin redirect Test timed out
